# ApiFinal/DTOs/UserLoginDto.cs
using System.ComponentModel.DataAnnotations;

namespace ApiEmpresa.DTOs
{
    public class UserLoginDto
    {
        [Required]
        public string Email { get; set; } = "";
        [Required]
        public string Password { get; set; } = "";
    }
}